+++ app/Http/Controllers/V1/ImageManipulationController.php
@@ -7,6 +7,7 @@ use App\Http\Resources\V1\ImageManipulationResource;
 use App\Models\Album;
 use App\Models\ImageManipulation;
 use App\Http\Requests\ResizeImageRequest;
+use Illuminate\Http\Request;
 use Illuminate\Http\UploadedFile;
 use Illuminate\Support\Facades\File;
 use Illuminate\Support\Str;
@@ -14,43 +15,37 @@ use Intervention\Image\Facades\Image;
 
 class ImageManipulationController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function index()
+    public function index(Request $request)
     {
-        return ImageManipulationResource::collection(ImageManipulation::paginate());
+        return ImageManipulationResource::collection(ImageManipulation::where('user_id',$request->user()->id)->paginate());
     }
 
-    public function byAlbum(Album $album)
+    public function byAlbum(Request $request, Album $album)
     {
+        if ($request->user()->id != $album->user_id){
+            return abort(403,'Unauthorized');
+        }
         $where = [
             'album_id' => $album->id
         ];
         return ImageManipulationResource::collection(ImageManipulation::where($where)->paginate());
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param \App\Http\Requests\ResizeImageRequest $request
-     * @return \Illuminate\Http\Response
-     */
     public function resize(ResizeImageRequest $request)
     {
         $all = $request->all();
-        /** @var UploadedFile|string $image */
         $image = $all['image'];
         unset($all['image']);
         $data = [
             'type' => ImageManipulation::TYPE_SIZE,
             'data' => json_encode($all),
-            'user_id' => null,
+            'user_id' => $request->user()->id,
         ];
         if (isset($all['album_id'])) {
-            //TODO
+            $album = Album::find($all['album_id']);
+            if ($request->user()->id != $album->user_id){
+                return abort(403,'Unauthorized');
+            }
             $data['album_id'] = $all['album_id'];
         }
         $dir = 'images/' . Str::random() . '/';
@@ -85,28 +80,23 @@ class ImageManipulationController extends Controller
         $image->resize($width, $height)->save($absolutePath . $resizedFilename);
         $data['output_path'] = $dir . $resizedFilename;
         $imageManipulation = ImageManipulation::create($data);
+
         return new ImageManipulationResource($imageManipulation);
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param \App\Models\ImageManipulation $image
-     * @return \Illuminate\Http\Response
-     */
-    public function show(ImageManipulation $image)
+    public function show(Request $request, ImageManipulation $image)
     {
+        if ($request->user()->id != $image->user_id){
+            return abort(403,'Unauthorized');
+        }
         return new ImageManipulationResource($image);
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param \App\Models\ImageManipulation $imageManipulation
-     * @return \Illuminate\Http\Response
-     */
-    public function destroy(ImageManipulation $image)
+    public function destroy(Request $request, ImageManipulation $image)
     {
+        if ($request->user()->id != $image->user_id){
+            return abort(403,'Unauthorized');
+        }
         $image->delete();
         return response('Delete Successfully', 204);
     }
